Add support for the composite overlay window.
Add support for using the composite overlay window for
GL output instead of rendering to the root window.  This is lets
compiz run on AIGLX without the _COMPIZ_GL_INCLUDE_INFERIORS hack that
we used to have.  The patch defaults to using the root window but adds
a '--use-cow' command line option to make compiz use the overlay
window.