Fix finite freeze w/ superkaramba after monitor standby.
Some system monitor applications like superkaramba widgets
(create and) destroy an (invisible) window every second, which
was triggering an unnecessary addWindowDamage in Animation
plugin for those windows. When the monitor is in stand-by mode,
compiz doesn't redraw the screen, causing these (potential)
closing animations to be delayed until the user returns
leading to high cpu usage or a (finite) freeze at that time.
This fix prevents the problem by not calling addWindowDamage
for such windows.