Ensure that plugins get correctly loaded after composite and opengl.
Some plugins had an oversight which caused them to not load correctly
due to a missing <deps> tag. Also some plugins should have been loading
after these plugins.

This fixes a potential glitch in newer versions of libdri.

It also fixes ugliness in the mass redirection/unredirection of windows
when the opengl plugin is loaded/reloaded