Fix startup order.
Previously we were doing something like this:

init -> screen init -> init core plugin screen -> create windows ->
eventLoop -> updatePlugins -> initPlugin (initScreen (initWindow))).

This order has a number of problems with it - namely that config plugins
will be loaded on one full pass of the event loop and then core options
are initialized, before which we were doing certain startup procedures
on the default options, which resulted in strange race condition bugs.
Also, if we are compositing we have a brief (visible) switch from 2D mode
to 3D mode - in reality we want to be in 3D mode ASAP.

The new order is init -> updatePlugins (load plugins specified on command
 line) -> screen init -> load command line plugins -> update plugins (load
new plugins specified by config plugins on the command line) -> init screens
-> create window (init plugins for window) -> eventLoop.

This means that it takes longer for us to get to the event loop, but at least
we can zip through the initialization phase quite quickly