Allow different font types in frames.
Currently we used a single PangoContext and PangoFontDescription for
everything ... this resulted in custom frames having the same titlebar
size which wasn't any good. This commit changes a lot, namely:

 * Frame is now linked to the decoration object itself
 * Moved style windows into the frame
 * Moved font context and font description into the frame
 * Added API for getting the titlebar font size: obviously with the
cairo decorations there isn't supposed to be any kind of visual
difference here but with the metacity decorations we need to get that
directly from libmetacity.

It should be noted that these commits are in need of a cleanup ... which
is coming in the next few commits.