Fix slow-motion-animations regression, most noticeable on slow machines or systems under heavy load. In that case animations were slowing down, instead of the old (correct) behaviour of simply skipping frames in the animation to ensure its duration is constant.
Ideally this hacking of timeDiff should be removed completely in future.
But before that is possible, all offending plugins would need to be fixed to
ensure their preparePaint functions ignore the msSinceLastPaint parameter on
the first frame. Since msSinceLastPaint can potentially be huge if the screen
has been idle for a while. Animations should not ever use msSinceLastPaint
until their second frame.