Fix crash in glDrawArrays
The NVIDIA driver looks to see if the legacy GL_VERTEX_ARRAY,
GL_TEXTURE_COORDINATES_ARRAY and other such client states are enabled
first before checking if a vertex buffer is bound and will prefer the
client buffers over the the vertex buffer object.

There isn't anywhere in the OpenGL specification that says that
the driver can't do that, however, it can inadvertently affect performance.

In addition, if no glVertexPointer was specified and the GL_VERTEX_ARRAY
client state is left on then it will attempt to read from NULL blindly.

The solution is to remove the legacy code that turned GL_VERTEX_ARRAY
and GL_TEXTURE_COORDINATES_ARRAY on by default

(LP :#1006211)