Added compiz_discover_tests.
gtest_add_tests is good for the basic Google Test usage, but its source
based scanning falls short in a couple of areas - namely it requires that
there be no space between TEST* and ( and it also does not work for templated
tests.

compiz_discover_tests builds a small C++ program to parse the output
of --gtest_list_tests and adds those at make-time to the CTestTestfile.cmake
in the current binary directory. This allows us to correctly introspect the test
binary and get /all/ the tests regardless of the source file.