Some decoration pixmaps get leaked on window resize, due to race conditions between gtk-window-decorator and the decor plugin.
That's a complex problem which requires a better interprocess communication
and resource ownership design. So I'm not trying to fix that directly.

Instead, what this does is limits pixmap leaks to the lifetime of the frame
window. So if any leaks do occur, they will be cleaned up and freed next
time the window is minimized or closed. Or to garbage collect for all windows,
just hit Ctrl+Alt+D twice.

(LP: #1057263)