Thumbnail plugin, new features:
Added a rounded background to the window title text.

Made the distance between text and thumbnail configurable.

Window-like-background or glow are now rendered around
the thumbnail live preview only, excluding the text.
This ensures no transparent hole is generated anymore and
fixes the look of the window title text rendering, while
enhancing the possibility for customization.

Code cleanup:

Return ASAP if the text plugin is not loaded.
No need to check if (t->text) twice.