Window: get proper outputDevice when the window is in a secondary workspace
When a window is in a secondary workspace, its serverGeometry points to an
offscreen location, then we can't just use it for getting the outputDevice as this
would fallback to the first output (as it's outside the actual screen area).
Then, in this case we need to subtract the window viewport coords.