blur: Port the blur plugin to use Modern OpenGL and enable it
The whole plugin needed to be re-written to a large extent in order to
use GLSL-based shaders instead of the old ASM based ones, use the new
GLFramebufferObject API and also play nice with damage tracking.

There's still a lot of work to do as the code is very messy.

Notably, generating mipmaps of FBOs seems to be an edge case which is
broken on a lot of hardware.