opengl: Add clearShaders method to GLWindow
This function is called by glDrawTexture once textures have been drawn. It
is responsible for clearing out the shader list for that window so that
the shader-list can be re-populated.

If a plugin wants to override glDrawTexture completely, it needs to
call clearShaders once it is done with the window texture.